[DO NOT REVIEW] Sync mongodb (#34)
* Spotless code style updates

* Tests - simplify and limit getNextResults looping

* Change the type of 'connection.uri' PASSWORD

Prevent username / password leak in logs.

KAFKA-378

* Removed legacy partition mapping

As it used the raw connection string it could leak credentials to the
log.  As changing connection string uri to password changed its output
there was no point in retaining it.

KAFKA-378

* Revert "Tests - simplify and limit getNextResults looping"

This reverts commit e2b640b9f04038187a9e2a1d1d1cdf29356110d3.

* Made Debezium (DBZ) ddl events a noop (#139)

KAFKA-274

* Improve "Could not convert value ..." messages (#140)

A viewer should not be left guessing whether the value was a null reference, or a "null" `CharSequence`.

* Add regular expressions in topic mapping (#141)

* Mark the cursor as nullable (#143)

Prevent NPE if the cursor cannot be recreated

KAFKA-383

* Added UpdateOneDefaultStrategy (#144)

A WriteStrategy that performs a `$set` upsert operation.
For scenarios where the topic data contains fragments of the
desired data in MongoDB

KAFKA-255

* Source: Added configuration to use the document key for the sourceRecord key

This is potentially a breaking change as the newly added configuration
`change.stream.document.key.as.key` defaults to true.

Previously, the resume token was used as the source key, but
it limits the usefulness of tombstones both for topic compactions
and for downstream implementations. 

Not all events relate to documents (eg drop collection) so fallbacks to 
resume token for those events.

As such this is considered both an improvement and a bug fix.

Set to false to revert back to the previous behaviour.

KAFKA-360

Co-authored-by: Ross Lawley <ross@mongodb.com>
Co-authored-by: Goncalo Pinho <goncalopinho@hotmail.com>

* Version: bump 1.11.0

* Version: bump 1.12.0-SNAPSHOT

* Fix the formatting of the `topic.namespace.map` documentation

* Mark nested Junit5 tests with `@Nested`

* Logging incompatible properties no longer NPEs on null values (#147)

KAFKA-390

Co-authored-by: Valentin Kovalenko <valentin.kovalenko@mongodb.com>

* Add wildcard support for starts with field name matching

KAFKA-391

* Version: 1.11.1

* Version: 1.12.0-SNAPSHOT

---------

Co-authored-by: Ross Lawley <ross.lawley@gmail.com>
Co-authored-by: Jinxin Liu <liukrimhim@gmail.com>
Co-authored-by: Ross Lawley <ross@mongodb.com>
Co-authored-by: Valentin Kovalenko <valentin.kovalenko@mongodb.com>
Co-authored-by: Goncalo Pinho <goncalopinho@hotmail.com>